Refactor: UI(페이지)와 비즈니스 로직(커스텀 훅) 분리
