--- src/features/chatbot/hooks/useChatInitialization.ts
@@ -1,3 +1,0 @@
-const useChatInitialization = () => {};
-
-export default useChatInitialization;
